arm: cluster_plug hotplug driver
Cluster-plug is a simple cluster-based hotplug implementation for homogeneous
ARM big.LITTLE systems such as the MSM8939. It operated based on the principle
that the big cluster is normally preferred, since it as as efficient as the
little one, but can go faster for single core performance. This hotplug will
only activate the little cluster for highly parallel loads, and then deactivate
it once the parallel load ends.

Change-Id: I188863d58285f363ac550528e664a9cad7ae1e5f
Signed-off-by: khanfrd <khanfrd94@gmail.com>