cluster_plug: Define a cluster as one little CPU and one big CPU
On MSM8996, the primary CPU is part of the little cluster, and unplugging
the primary CPU causes chaos. Instead of splitting clusters up into a set
of little cores and a set of big cores, split them up into one little core
and one big core each. That way, we can still have a CPU0 without sacrificing
performance and efficiency.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: khanfrd <khanfrd94@gmail.com>